chore: rename account.domain
