+++ frontend/core/stores/account/spec.d.ts
@@ -3,6 +3,7 @@ import type { TAccount, TCommunity, TUser } from '~/spec'
 export type TStore = {
   user: TUser | null
   isLogin: boolean
+  isModerator: boolean
   userSubscribedCommunities: TCommunity[] | null
 
   // views
