+++ frontend/core/stores/provider.landing.tsx
@@ -1,7 +1,7 @@
 import type { FC } from 'react'
 import METRIC from '~/const/metric'
 import type { TCommunity } from '~/spec'
-import AccountStoreProvider from '~/stores/account.domain/provider'
+import AccountStoreProvider from '~/stores/account/provider'
 import CommunityStoreProvider from '~/stores/community/provider'
 import DashboardStoreProvider from '~/stores/dashboard/provider'
 import GeneralStoreProvider from '~/stores/general/provider'
