+++ frontend/core/stores/provider.tsx
@@ -1,6 +1,6 @@
 import type { FC } from 'react'
 import type { TCommunity } from '~/spec'
-import AccountStoreProvider from '~/stores/account.domain/provider'
+import AccountStoreProvider from '~/stores/account/provider'
 import CommunityStoreProvider from '~/stores/community/provider'
 import DashboardStoreProvider from '~/stores/dashboard/provider'
 import GeneralStoreProvider from '~/stores/general/provider'
